requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
chromadb>=0.4.0
pypdf>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
plotly>=5.17.0
pandas>=2.0.0
numpy>=1.24.0